Feature: numero de alunos pode ser limitado

+++ sme_terceirizadas/kit_lanche/models.py
@@ -147,7 +147,7 @@ class SolicitacaoKitLancheAvulsaBase(TemChaveExterna,  # type: ignore
 
 
 class SolicitacaoKitLancheAvulsa(ExportModelOperationsMixin('kit_lanche_avulsa'), SolicitacaoKitLancheAvulsaBase):
-    quantidade_alunos = models.PositiveSmallIntegerField(blank=True, null=True)
+    quantidade_alunos = models.BigIntegerField(blank=True, null=True)
     escola = models.ForeignKey('escola.Escola', on_delete=models.DO_NOTHING,
                                related_name='solicitacoes_kit_lanche_avulsa')
     alunos_com_dieta_especial_participantes = models.ManyToManyField('escola.Aluno')
